moved templates to common place
